requirements.txt
streamlit
pymupdf
openai
azure-search-documents
azure-core
python-dotenv
